[Ref] Silence 2 64bit warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4065 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/misc_util.cpp
@@ -417,7 +417,7 @@ std::wstring GUIDToString(GUID guid)
 //----------------------------------
 {
 	std::vector<OLECHAR> tmp(256);
-	::StringFromGUID2(guid, &tmp[0], tmp.size());
+	::StringFromGUID2(guid, &tmp[0], static_cast<int>(tmp.size()));
 	return &tmp[0];
 }
